Remove OCR2VRF skeleton reliance on chainlink-relay types
