+++ core/services/relay/evm/evm.go
@@ -191,66 +191,3 @@ func (p *medianProvider) ReportCodec() median.ReportCodec {
 func (p *medianProvider) MedianContract() median.MedianContract {
 	return p.medianContract
 }
-
-// Relayer with added DKG and OCR2VRF provider functions.
-type ocr2vrfRelayer struct {
-	*Relayer
-}
-
-var _ relaytypes.OCR2VRFRelayer = (*ocr2vrfRelayer)(nil)
-
-func NewOCR2VRFRelayer(relayer interface{}) relaytypes.OCR2VRFRelayer {
-	return &ocr2vrfRelayer{relayer.(*Relayer)}
-}
-
-type dkgProvider struct {
-	*configWatcher
-	contractTransmitter *ContractTransmitter
-}
-
-var _ relaytypes.DKGProvider = (*dkgProvider)(nil)
-
-func (r *ocr2vrfRelayer) NewDKGProvider(rargs relaytypes.RelayArgs, transmitterID string) (relaytypes.DKGProvider, error) {
-	configWatcher, err := newConfigProvider(r.lggr, r.chainSet, rargs)
-	if err != nil {
-		return nil, err
-	}
-	contractTransmitter, err := newContractTransmitter(r.lggr, rargs, transmitterID, configWatcher)
-	if err != nil {
-		return nil, err
-	}
-	return &dkgProvider{
-		configWatcher:       configWatcher,
-		contractTransmitter: contractTransmitter,
-	}, nil
-}
-
-func (c *dkgProvider) ContractTransmitter() types.ContractTransmitter {
-	return c.contractTransmitter
-}
-
-type vrfProvider struct {
-	*configWatcher
-	contractTransmitter *ContractTransmitter
-}
-
-var _ relaytypes.DKGProvider = (*dkgProvider)(nil)
-
-func (r *ocr2vrfRelayer) NewOCR2VRFProvider(rargs relaytypes.RelayArgs, transmitterID string) (relaytypes.OCR2VRFProvider, error) {
-	configWatcher, err := newConfigProvider(r.lggr, r.chainSet, rargs)
-	if err != nil {
-		return nil, err
-	}
-	contractTransmitter, err := newContractTransmitter(r.lggr, rargs, transmitterID, configWatcher)
-	if err != nil {
-		return nil, err
-	}
-	return &dkgProvider{
-		configWatcher:       configWatcher,
-		contractTransmitter: contractTransmitter,
-	}, nil
-}
-
-func (c *vrfProvider) ContractTransmitter() types.ContractTransmitter {
-	return c.contractTransmitter
-}
